requirements.txt
accelerate
pydub
nnAudio
PyYAML
transformers
hydra-core
tensorboard